add different warehouse
