+++ fe/fe-core/src/main/java/com/starrocks/warehouse/Cluster.java
@@ -24,6 +24,8 @@ import com.starrocks.persist.gson.GsonUtils;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class Cluster implements Writable {
@@ -31,21 +33,30 @@ public class Cluster implements Writable {
     private long id;
     @SerializedName(value = "wgid")
     private long workerGroupId;
+    @SerializedName(value = "cnIds")
+    private Set<Long> cnIds = new HashSet<>();
 
     // Note: we only record running sqls number and pending sqls in Warehouse and Cluster
     // We suppose that sql queue tool has nothing to do with  Cluster,
     // Cluster offers related interfaces and sql queue tool will update counter according to the implementation of sqls.
     private AtomicInteger numRunningSqls;
 
-    public Cluster(long id, long workerGroupId) {
+    public Cluster(long id) {
         this.id = id;
+    }
+
+    public Cluster(long id, long workerGroupId) {
+        this(id);
         this.workerGroupId = workerGroupId;
     }
 
-    // set the associated worker group id when resizing
-    /*    public void setWorkerGroupId(long id) {
-        this.workerGroupId = id;
-    }*/
+    public void addNodes() {
+
+    }
+
+    public void clearComputeNodes() {
+        cnIds.clear();
+    }
 
     public long getWorkerGroupId() {
         return workerGroupId;
@@ -61,12 +72,6 @@ public class Cluster implements Writable {
     public int getPendingSqls() {
         return -1;
     }
-    /*    public int setRunningSqls(int val) {
-        return 1;
-    }
-    public int addAndGetRunningSqls(int delta) {
-        return 1;
-    }*/
 
     public void getProcNodeData(BaseProcResult result) {
         result.addRow(Lists.newArrayList(String.valueOf(this.getId()),
